Fix Docker setup reachability and .env location (#87)
The guided setup was only available as the standalone `kptncook-setup`
console script, which is unreachable through the Docker image's
`kptncook` entrypoint. Expose it as a `kptncook setup` subcommand and
point the missing-config message at it.

Move the setup implementation into `kptncook.setup` so the main CLI can
register it without importing `kptncook_setup` (which would create an
import cycle that breaks the standalone entrypoint). `kptncook_setup`
now re-exports from `kptncook.setup`.

Also honor `KPTNCOOK_HOME` when locating `.env` so the config lives in
the same directory as the rest of kptncook's data. In Docker
(`KPTNCOOK_HOME=/data`) the `.env` now persists in the mounted volume
instead of an ephemeral `~/.kptncook/.env`.
